Improve parameters, split cubits in different files
