+++ app/lib/cubits/editor_view.dart
@@ -1,0 +1,82 @@
+part of 'editor_runtime.dart';
+
+@freezed
+sealed class EditorViewState with _$EditorViewState {
+  const factory EditorViewState({
+    @Default(UtilitiesState()) UtilitiesState utilities,
+    @Default(ViewOption()) ViewOption viewOption,
+    @Default(true) bool areaNavigatorCreate,
+    @Default(true) bool areaNavigatorExact,
+    @Default(false) bool areaNavigatorAsk,
+    @Default(false) bool navigatorEnabled,
+    @Default(NavigatorPage.waypoints) NavigatorPage navigatorPage,
+    @Default('') String userName,
+  }) = _EditorViewState;
+}
+
+class EditorViewCubit extends Cubit<EditorViewState> {
+  EditorViewCubit({this.editorSessionCubit, EditorViewState? initial})
+    : super(initial ?? const EditorViewState());
+
+  final EditorSessionCubit? editorSessionCubit;
+  ToolCubit? _toolCubit;
+
+  void bindToolCubit(ToolCubit toolCubit) {
+    _toolCubit = toolCubit;
+  }
+
+  @override
+  void onChange(Change<EditorViewState> change) {
+    super.onChange(change);
+    if (change.nextState.userName != change.currentState.userName) {
+      _toolCubit?.scheduleNetworkingState();
+    }
+  }
+
+  void replace(EditorViewState state) => emit(state);
+
+  void updateUtilities({UtilitiesState? utilities, ViewOption? view}) {
+    emit(
+      state.copyWith(
+        utilities: utilities ?? state.utilities,
+        viewOption: view ?? state.viewOption,
+      ),
+    );
+    if (utilities != null) {
+      editorSessionCubit?.updateUtilities(utilities);
+    }
+  }
+
+  void setAreaNavigator({bool? create, bool? exact, bool? ask}) {
+    emit(
+      state.copyWith(
+        areaNavigatorCreate: create ?? state.areaNavigatorCreate,
+        areaNavigatorExact: exact ?? state.areaNavigatorExact,
+        areaNavigatorAsk: ask ?? state.areaNavigatorAsk,
+      ),
+    );
+    editorSessionCubit?.updateAreaNavigator(
+      create: create,
+      exact: exact,
+      ask: ask,
+    );
+  }
+
+  void setNavigator({bool? enabled, NavigatorPage? page}) {
+    emit(
+      state.copyWith(
+        navigatorEnabled: enabled ?? state.navigatorEnabled,
+        navigatorPage: page ?? state.navigatorPage,
+      ),
+    );
+    editorSessionCubit?.updateNavigator(enabled: enabled, page: page);
+  }
+
+  void setUserName(String name) => emit(state.copyWith(userName: name));
+
+  @override
+  Future<void> close() {
+    _toolCubit = null;
+    return super.close();
+  }
+}
